Resolves the conflicts blocking PR 326. The branch was 12 ahead and 91 behind;
five conflicts, of which one carried a real decision.

**QrResolver.cs - BOTH sides, not one.** This is the trap that bit the last
merge, so it was checked rather than resolved by file. The branch had changed
WHICH name wins (the profile's, because the account may not exist at all for a
walk-in), while main had ADDED two constructor arguments - BadgeBatchId and
EditionYear. Taking the branch's side would have dropped them; taking main's
would have undone the name decision. BadgeBatchId is what the badge self-claim
guard reads to tell a bulk-order badge from a walk-in's, and EditionYear is the
only expiry a minted QR has, so dropping them is not cosmetic. The projection
already selects both, so the merged call compiles as written.

**App migration - regenerated, not hand-merged.** Both sides had regenerated
their App InitialCreate (branch 20260814134021, main 20260814115348), so git saw
a rename/rename. Every competing migration and the App snapshot were deleted and
one regenerated from the merged model. Verified on the result: the branch's file
keys are present (ImageFileId, ThumbnailFileId, RecordingFileId, LogoFileId,
PhotoFileId), main's are present (EditionYear, AdmissionState, BadgeBatchId, the
EventEdition table), and exactly one InitialCreate per context remains - which is
what D-895's SchemaFreezeTests, arriving with this merge, now enforces. Identity
needed no regeneration: the branch never touched it.

CLAUDE.md's freeze table is re-pointed to the regenerated App id in this same
commit, so the governing doc moves WITH the schema rather than after it. The
D-895 carve-out is NARROWED rather than closed: the two archive pointers below
are still strings and that conversion is in flight, so declaring it spent would
put a half-finished pipeline behind an owner lift. It now closes on evidence
instead - when the ratchet's countdown empties, the carve-out is spent.

The rest were not decisions: DECISIONS_LOG took main's block after checking it
is a strict superset (nothing existed only on the branch); SimfUser.cs was a
comment re-wording around identical code; home_screen.dart was two DIFFERENT
imports, so both are kept.

**One pre-existing red fixed.** MediaPointerRatchetTests'
A_recorded_remainder_that_is_gone_is_deleted was already failing at the branch
tip: the branch's own last commit converted six stream/video columns and did not
shorten KnownRemaining with them, which is precisely what that test exists to
catch. The six are removed; ArchivePastSpeaker.PhotoRelativePath and
ArchiveMediaItem.Url stay, because those really are still strings and are the
work in flight. Confirmed pre-existing rather than merge-induced by reading the
branch tip - the properties are gone there while the list still names them, and
main never had the conversion at all.

Build 0 warnings / 0 errors. Domain 23, Application 129, ControlPanel 651,
Web 142, ApiClient 50.

+++ src/Mobile/simf_app/lib/features/venuemap/venue_map_screen.dart
@@ -131,9 +131,8 @@ class _VenueMapScreenState extends ConsumerState<VenueMapScreen> {
   Size get _viewport => context.size ?? Size.zero;
 
   void _zoomBy(double factor) {
-    final scale = (_transform.value.getMaxScaleOnAxis() * factor)
-        .clamp(0.3, 4.0)
-        .toDouble();
+    final scale =
+        (_transform.value.getMaxScaleOnAxis() * factor).clamp(0.3, 4.0);
     final centre = _viewport.center(Offset.zero);
     // Re-anchor the zoom on the viewport centre.
     final scenePoint = _transform.toScene(centre);
